Change project name to QueryDuck and split repository
Repository changes to coincide with the name change:

- Split several parts to new repositories, see below
- Change most occurences of "CrunchyVicar" to "QueryDuck"
- Change "CV" in class names to "QD".
- Rename "common/crunchylib" to "queryduck"
- Update history, gitignore
- Update .gitignore

The orignal repository has been split as follows:

- "cliclient" directory split to "queryduck-cli" repository
- "server" directory split to "queryduck-server" repository
- "webclient" directory split to "queryduck-web" repository
- Main repository remains with all other files, but is now called "queryduck"

Commit history has been preserved, with a few caveats:

Commit messages in the original repository were prefixed with the name of the
main subdirectory. I have removed this prefix in the split repositories, but
retained it in the main repository for the remaining commits. The prefix will
no longer be used for future commits.

Since files were moved between directories in the original repository on
several occasions, it will appear in the new structure as though the files
disappeared from one repository and reappeared in the other. This will
probably make the history a bit confusing. It should only apply to commits
older than this one, though.