Split off functionality from Result into Collection
Result will still contain the primary results of a query, but additional
Statements and Files are now part of a separate Collection object.

Repository,query() will now return a tuple of Result and Collection.

While slightly more verbose, this makes sense as the two sets of data are not
directly connected (but still returned in one server call for efficiency).