Create and use GroupedCollection class
This class has a similar interface to a regular Collection, but contains other
Collections rather than Statements. This allows additional results to be
appended as they are fetched.

There is no additional indexing beyond what happens in the child Collections,
so it won't scale well for large numbers of child Collections. Nesting should
be possible but should probably be avoided for the same reason.